CLN: re-arrange tests
